Change IOStream interface and implementations to support large files

+++ taglib/toolkit/tbytevectorstream.h
@@ -109,7 +109,7 @@ namespace TagLib {
      *
      * \see Position
      */
-    void seek(long offset, Position p = Beginning);
+    void seek(long long offset, Position p = Beginning);
 
     /*!
      * Reset the end-of-file and error flags on the file.
@@ -119,12 +119,12 @@ namespace TagLib {
     /*!
      * Returns the current offset within the file.
      */
-    long tell() const;
+    long long tell() const;
 
     /*!
      * Returns the length of the file.
      */
-    long length();
+    long long length();
 
     /*!
      * Truncates the file to a \a length.
